api-verifier: add cloudbuild.yaml + update Dockerfile for GCB
This adds a cloudbuild.yaml file and updates the Dockerfile so that we
can build this using Google Container Builder.

The cloudbuild.yaml file produces an image with two tags: `latest` and a
name that's passed in by the user via a command line flag to `gcloud
container builds submit`.

I'm going to be making these updates for all the containers, but I
wanted to do this one first because it's one of the simpler ones (it
doesn't need any additional steps).