fluentd-logger: add cloudbuild.yaml + update Dockerfile for GCB
This adds a cloudbuild.yaml file and updates the Dockerfile so that we
can build this using Google Container Builder.

The cloudbuild.yaml file produces an image with two tags: latest and a
name that's passed in by the user via a command line flag to gcloud
container builds submit.

See also #101.