cloudsql-proxy: add cloudbuild.yaml file for GCR
This adds a cloudbuild.yaml file and so that we can build this using
Google Container Builder.

The cloudbuild.yaml file produces an image with two tags: latest and a
name that's passed in by the user via a command line flag to gcloud
container builds submit.

It also builds the go binary used by the Docker container, which is
pinned to a specific commit. This replicates the logic in the ./build.sh
file.

See also #101.