memcache-proxy: add cloudbuild.yaml + custom Dockerfile.builder for GCR (#107)
This adds a cloudbuild.yaml config file as well as a custom Dockerfile
to specify the build environment for this sidecar so that we can build
it with Google Container Builder.

The cloudbuild.yaml file produces an image with two tags: latest and a
name that's passed in by the user via a command line flag to gcloud
container builds submit. It also builds a special container used to run
the `build.sh` script, but this container is not saved externally.

Also, I updated the `build.sh` script to use `mktemp -d` instead of
`--directory` because the version of `mktemp` that container builder
uses doesn't support `--directory`. This should be backwards compatible.