Fix key fetching time randomization (#119)
The previous code applied randomization at build time; the new code applies randomization at container run time. Also, the now uses a bash script as the entrypoint. This allows all key fetching and cron job setup to be removed from the Python code and done natively as shell commands, which is easier to read and understand.