Add dockerstats OpenTelemetry receiver.
This receiver periodically scrapes docker API for container stats and
exports them as metrics. The change only adds memory usage/limit as
metris, but would be extended to support more.

This change also upgrades the OpenTelemetry collector library versions
to v0.6.0. There were a lot of breaking API changes since the
OpenTelemetry monitoring API is still in Alpha.

Additionally, we run `go test` as an additional step in Cloud Build.