Add dockerstats receiver to OpenTelemetry collector (#143)
* Add dockerstats OpenTelemetry receiver.

This receiver periodically scrapes docker API for container stats and
exports them as metrics. The change only adds memory usage/limit as
metris, but would be extended to support more.

This change also upgrades the OpenTelemetry collector library versions
to v0.6.0. There were a lot of breaking API changes since the
OpenTelemetry monitoring API is still in Alpha.

Additionally, we run `go test` as an additional step in Cloud Build.

* Add tests for dockerstats receiver.

* Run gofmt on scraper_test

* dockerstats code review fixes:

* Add comment about docker naming
* Add info log when starting a new export cycle

Co-authored-by: Sanrag Sood <sanragsood@google.com>